services/ingestor/requirements.txt
PyYAML
psycopg2-binary
openai
azure-storage-blob
